Massive revamp to game list
Small QOL changes and an entire revamp to the games list. Removed aquapark slides because cringe. Bolded the text on the footer copyright bar. Added game icon folder for the games list. Changed the fetch for ludos icons to be directly to the file rather than from the github page.